requirements.txt
fastapi
uvicorn
pydantic
google-genai
google-cloud-aiplatform
vertexai
